Let's wait to do Clevo support for now.  Started to rewrite Razer from scratch
