+++ Project-Aurora/Project-Aurora/Devices/DeviceManager.cs
@@ -101,11 +101,12 @@ namespace Aurora.Devices
             devices.Add(new DeviceContainer(new Devices.Logitech.LogitechDevice()));         // Logitech Device
             devices.Add(new DeviceContainer(new Devices.SteelSeries.SteelSeriesDevice()));   // SteelSeries Device
             devices.Add(new DeviceContainer(new Devices.Wooting.WootingDevice()));           // Wooting Device
-            devices.Add(new DeviceContainer(new Devices.Clevo.ClevoDevice()));               // Clevo Device
+            devices.Add(new DeviceContainer(new Devices.Razer.RazerDevice()));               // Razer Device
             /*
             devices.Add(new DeviceContainer(new Devices.Corsair.CorsairDevice()));           // Corsair Device
-            devices.Add(new DeviceContainer(new Devices.Razer.RazerDevice()));               // Razer Device
+
             devices.Add(new DeviceContainer(new Devices.Roccat.RoccatDevice()));             // Roccat Device
+            devices.Add(new DeviceContainer(new Devices.Clevo.ClevoDevice()));               // Clevo Device
             devices.Add(new DeviceContainer(new Devices.CoolerMaster.CoolerMasterDevice())); // CoolerMaster Device
             devices.Add(new DeviceContainer(new Devices.AtmoOrbDevice.AtmoOrbDevice()));     // AtmoOrb Ambilight Device
             devices.Add(new DeviceContainer(new Devices.UnifiedHID.UnifiedHIDDevice()));     // UnifiedHID Device
